+++ src/main/java/net/meander/subtlyd/mixin/client/renderer/entity/ZombifiedPiglinRendererMixin.java
@@ -1,6 +1,7 @@
 package net.meander.subtlyd.mixin.client.renderer.entity;
 
 import net.meander.subtlyd.client.renderer.UndeadRenderStateAccessor;
+import net.meander.subtlyd.network.syncher.SynchedEntityDataSD;
 import net.meander.subtlyd.util.Util;
 import net.minecraft.client.renderer.entity.ZombifiedPiglinRenderer;
 import net.minecraft.client.renderer.entity.state.ZombifiedPiglinRenderState;
@@ -15,6 +16,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(ZombifiedPiglinRenderer.class)
 public class ZombifiedPiglinRendererMixin {
     private final Identifier ZOMBIFIED_PIGLIN_LEADER_LOCATION = Util.identifier("textures/entity/piglin/zombified_piglin_leader.png");
+    private final Identifier BABY_ZOMBIFIED_PIGLIN_LEADER_LOCATION = Util.identifier("textures/entity/piglin/baby_zombified_piglin_leader.png");
 
     /**
      * Changes the zombified piglin leader texture to their unique design.
@@ -26,7 +28,7 @@ public class ZombifiedPiglinRendererMixin {
         Identifier leaderLocation = cir.getReturnValue();
 
         if (((UndeadRenderStateAccessor) state).subtlyDungeons$isLeader()) {
-                leaderLocation = ZOMBIFIED_PIGLIN_LEADER_LOCATION;
+                leaderLocation = state.isBaby ? BABY_ZOMBIFIED_PIGLIN_LEADER_LOCATION : ZOMBIFIED_PIGLIN_LEADER_LOCATION;
         }
         cir.setReturnValue(leaderLocation);
     }
@@ -34,7 +36,7 @@ public class ZombifiedPiglinRendererMixin {
     @Inject(method = "extractRenderState(Lnet/minecraft/world/entity/monster/zombie/ZombifiedPiglin;Lnet/minecraft/client/renderer/entity/state/ZombifiedPiglinRenderState;F)V",
             at = @At("TAIL"))
     private void setLeaderRenderState(ZombifiedPiglin entity, ZombifiedPiglinRenderState state, float partialTicks, CallbackInfo ci) {
-        double ZOMBIE_BASE_HEALTH_POINTS = 20D;
-        ((UndeadRenderStateAccessor) state).subtlyDungeons$setLeader(entity.getMaxHealth() > ZOMBIE_BASE_HEALTH_POINTS);
+        boolean isLeader = entity.getEntityData().get(SynchedEntityDataSD.DATA_ID_ZOMBIE_LEADER);
+        ((UndeadRenderStateAccessor) state).subtlyDungeons$setLeader(isLeader);
     }
 }